requirements.txt
python-dotenv-1.1.1
altair-5.5.0 
blinker-1.9.0 
cachetools-6.2.0 
click-8.2.1 
protobuf-6.32.0 
pyarrow-21.0.0 
pydeck-0.9.1 
streamlit-1.49.0 
tenacity-9.1.2 
toml-0.10.2 
watchdog-6.0.0
openai 1.102.0